requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
community==1.0.0b1
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
executing==2.0.1
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.4
ipykernel==6.29.4
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
packaging==24.1
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-louvain==0.16
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.0.3
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
stack-data==0.6.3
sympy==1.12.1
threadpoolctl==3.5.0
torch==2.2.2
torch-sparse==0.6.18
torch_geometric==2.5.3
torch_scatter==2.1.2
torchaudio==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.4
yarl==1.9.4
zipp==3.19.2
